--- packages/binary/valuetransfer/payload/cached_object.go
@@ -1,1 +1,0 @@
-package payload
